:art: Separated user authentication into its own blueprint called auth.
