+++ app/auth/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-main = Blueprint("main", __name__)
+auth = Blueprint("auth", __name__)
 
 from . import views
