+++ app/auth/forms.py
@@ -1,4 +1,3 @@
-import re
 from flask_wtf import FlaskForm
 from wtforms import BooleanField, StringField, PasswordField, SubmitField
 from wtforms.validators import InputRequired, EqualTo, Length, Email, ValidationError
